Fix infinite loop.
Use QQmlParserStatus to delay refresh() until the favorites model
prop binding is evaluated, otherwise AbstractModel::favoritesModel()
loops here due to rootModel() returning itself when the
RecentUsageModel is used standalone rather than as a node in the
RootModel tree.

This fixes things (and is nice for performance anyway), but clearly
the AbstractModel::rootModel() should get a fix for the underlying
issue.

CCMAIL:ivan.cukic@kde.org